@@ -1,6 +1,6 @@
 {
     "name": "myspot",
-    "version": "0.2.3",
+    "version": "0.3.0",
     "private": true,
     "scripts": {
         "start": "vite",
@@ -26,13 +26,13 @@
         "jotai": "^1.6.0",
         "prop-types": "^15.8.1",
         "query-string": "^7.1.1",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-copy-to-clipboard": "^5.0.4",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-ga": "^3.3.0",
         "react-intersection-observer": "^8.33.1",
         "react-kakao-maps-sdk": "^1.0.4",
-        "react-query": "^3.34.11",
+        "react-query": "^3.39.1",
         "react-query-helper": "^0.1.0",
         "react-responsive": "^9.0.0-beta.4",
         "react-router": "^6.2.1",
@@ -47,9 +47,9 @@
         "@testing-library/cypress": "^8.0.2",
         "@types/node": "^12.0.0",
         "@types/prop-types": "^15",
-        "@types/react": "^17.0.0",
+        "@types/react": "^18.0.15",
         "@types/react-copy-to-clipboard": "^5",
-        "@types/react-dom": "^17.0.0",
+        "@types/react-dom": "^18.0.6",
         "@typescript-eslint/eslint-plugin": "^5.2.0",
         "@typescript-eslint/parser": "^5.2.0",
         "@vitejs/plugin-react-refresh": "^1.3.6",
@@ -90,5 +90,9 @@
             "last 1 safari version"
         ]
     },
-    "packageManager": "yarn@3.1.1"
+    "packageManager": "yarn@3.1.1",
+    "resolutions": {
+        "@types/react": "17.0.2",
+        "@types/react-dom": "17.0.2"
+    }
 }
